KDEV-1796: Separate createAuthenticator from the User class
